fix: let daily production post its material issue
The inventory issue negated a bare parameter, which PostgreSQL cannot
type: every daily entry that reported material usage failed with
"operator is not unique: - unknown" and surfaced as a 500. The quantity
is now cast before it is negated.

Two problems found with it while running the SQL suite against a real
engine for the first time: a routing for a part of another company
answered 404 "planning not found" because the new row lock let the
no-rows error escape before validation, and JSON documents were passed
as bytes rather than text. Server errors now name the production module
instead of planning.

+++ backend/internal/production/inventory_bridge.go
@@ -20,8 +20,10 @@ func issueMaterialToInventory(ctx context.Context, tx database.TenantTx, a Actor
 		if !material.Quantity.IsPositive() {
 			continue
 		}
+		// The quantity is cast before it is negated: PostgreSQL cannot resolve
+		// unary minus on an untyped parameter.
 		if _, e := tx.Exec(ctx, inventoryLedgerInsert+`
- VALUES($1,'PRODUCTION_ISSUE',$2,-$3,$4,'RAW MATERIAL','PRODUCTION','PRODUCTION_ENTRY',$5,$6)`,
+ VALUES($1,'PRODUCTION_ISSUE',$2,-($3::numeric),$4,'RAW MATERIAL','PRODUCTION','PRODUCTION_ENTRY',$5,$6)`,
 			a.TenantID, material.MaterialID, material.Quantity, material.UnitCode, entryID, a.UserID); e != nil {
 			return e
 		}
